warning message when depositing in TON + forge fmt

+++ contracts/src/Mock/interfaces/Layer2I.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.4;
 
 interface Layer2I {
-  function operator() external view returns (address);
-  function isLayer2() external view returns (bool);
-  function currentFork() external view returns (uint);
-  function lastEpoch(uint forkNumber) external view returns (uint);
-  function changeOperator(address _operator) external;
+    function operator() external view returns (address);
+    function isLayer2() external view returns (bool);
+    function currentFork() external view returns (uint256);
+    function lastEpoch(uint256 forkNumber) external view returns (uint256);
+    function changeOperator(address _operator) external;
 }
